Support SSH agent for git authentication (#566)
* Remove make bookkeeping file

The file .deps is used in the Makefile to signal that the script to
download dependencies has been run. At some point it was added to the
git repo, meaning that the dependencies script will never run in a new
clone. Remove it.

Signed-off-by: Michael Bridgen <michael@weave.works>

* Support ssh-agent in wego app add

Many people using SSH will _also_ be running ssh-agent. This commit
changes the add command to first check whether ssh-agent is in use and
if so use it for authentication, and otherwise leaves the key-finding
logic undisturbed.

 - I translated the if..else if.. to a switch, which I find reads
   better after you have a couple of branches.

 - The private key file path is now defined locally, rather than
   mutating a field of the struct value `params`. The field isn't used
   by the package defining AddParams, and arguably shouldn't be
   there.

Signed-off-by: Michael Bridgen <michael@weave.works>

* Use ssh-agent and fall back to guessing key files

The prior commit added ssh-agent support by checking whether
SSH_AUTH_SOCK is set in the environment, and using go-git's SSH Agent
auth method if so. This works but it's not great, since it's all or
nothing -- if the user happens not to have added an appropriate key to
the agent, it will fail, and they will have to figure out why.

Instead, the usual approach (e.g., that taken by git) is to try
methods until one succeeds. In other words, try to get a key from
ssh-agent, and if that doesn't work, guess a key file and try using
that.

go-git's ssh package doesn't suppport that directly, but I can build
it from the bits there (and bits already here).

Signed-off-by: Michael Bridgen <michael@weave.works>

* retabify

* Support ssh-agent in wego app add

Many people using SSH will _also_ be running ssh-agent. This commit
changes the add command to first check whether ssh-agent is in use and
if so use it for authentication, and otherwise leaves the key-finding
logic undisturbed.

 - I translated the if..else if.. to a switch, which I find reads
   better after you have a couple of branches.

 - The private key file path is now defined locally, rather than
   mutating a field of the struct value `params`. The field isn't used
   by the package defining AddParams, and arguably shouldn't be
   there.

Signed-off-by: Michael Bridgen <michael@weave.works>

* Use ssh-agent and fall back to guessing key files

The prior commit added ssh-agent support by checking whether
SSH_AUTH_SOCK is set in the environment, and using go-git's SSH Agent
auth method if so. This works but it's not great, since it's all or
nothing -- if the user happens not to have added an appropriate key to
the agent, it will fail, and they will have to figure out why.

Instead, the usual approach (e.g., that taken by git) is to try
methods until one succeeds. In other words, try to get a key from
ssh-agent, and if that doesn't work, guess a key file and try using
that.

go-git's ssh package doesn't suppport that directly, but I can build
it from the bits there (and bits already here).

Signed-off-by: Michael Bridgen <michael@weave.works>

* Make function private that didn't need to be public

* force remove .deps in Makefile

* update based on review comments

Co-authored-by: Jerry Jackson <jerry@weave.works>