Restructure GitHub auth test to remove flakiness
As best as I can tell, the flakiness here was coming from two places:

1. Plain ol' data races, where we were sharing a variable between
   threads, updating it one thread and reading it in another. Sometimes
   these wouldn't happen in the right order, causing failures.
2. Race conditions when interacting with the `clock` library, where (for
   example) we would kick off a piece of code that we knew would call
   `Sleep`, call `runtime.Gosched` to give it a chance to execute, and
   then tell the `clock` library to advance time. It looks like these,
   again, weren't always happening in order.

A bunch of the complexity here stems from the fact that we're poking our
clock from one end with the `Sleep` function, then poking it from the
other end to tell it time has advanced. Since all we're relying on in
our non-test code is a `sleep` function, we can fake this out with a
small `struct`, then use the `mockRoundTripper` to keep track of what
the "current time" is when it receives requests.

Another change is using a channel to communicate when `pollAuthStatus`
has completed, so that we can wait for this (and know nothing else is
going to change) before testing the state of the system.

Finally, I took the opportunity to restructure the tests a little bit to
take advantage of Ginkgo's `Context`s, providing a slightly clearer
separation between what scenario we're testing and what behavior we want
to see.