Copy existing ui-run code to gitops-server
We want to remove the `ui run` command and create a separate binary
which will run the gitops ui. This (very brutishly) copies the existing
ui run code over the gitops-server code to be used as a separate binary.

The gitops-server is being over-written as it was apparently created as
a convenience for developing the UI. As part of this work that should
become obsolete as a pattern and be replaced by the actual UI binary