Use system flux instead of an embedded flux
This just calls "flux" instead of a specific path - it has the neat
side-effect that instead of using WEAVE_GITOPS_FLUX_BIN_PATH, you can
just set your own flux bin path using the PATH variable.

Right now, there are only one thing we do with this flux: create a
secret, in order to be able to access private git repos, which is
called by auth which is called by both upgrade and profiles.

There's still a flux installed by `make dependencies` - I've adapted
the auth tests to use the one installed by that, but as it's not added
to your path, it's not going to be used by the local CLI. If you want
to work on upgrade or profiles, you might need to install it on your
own. "Trusted Flux" is part of our fairly short-term plans, so this
manual step should be temporary.