Tidy up use of `make cmd/gitops/...` & `touch`
Some common patterns in the github workflows is to either create a fake
ui dist or flux file (`mkdir ... && touch ...`) or to run `npm ci` then
separately generate the ui. The former can now be removed as the
directories to be embedded have had keep files added while the latter
can be replaced with the simple `make ui` which performs all necessary
steps to generate the ui.