Fix a couple of docker target problems (#1545)
* Fix `make local-docker-image` after docker file rename

Should this target exist? Not sure. However, this target does exist,
is pulled in by the acceptance test target automatically, and it was
broken. So fix it.

What this target used to do differently from the default targets was
to tag the image with an internal name instead of a public one. Should
this distinction exist? Not sure. However, as both targets exist, I'll
restore the distinction, too.

* Fix which docker image needs to embed flux