Reformat go.mod
When I run `go mod tidy` and have 2 require blocks, the direct
dependencies are in the top block and the indirect dependencies are in
the bottom block, which makes it easier to understand our
dependencies, e.g. "why do we depend on 3 yaml libraries?". However,
we had 4 blocks, so go just leaves them as they are.