feat: List namespaces endpoint
Tried to return as many information as I we can list labels, annotations,
and status.

On the request, it does not accept any filtering options. If we want to
pass in filters, we can add when we need filtering on server side. Right
now list endpoints are returning everything without filters and the UI
does the filtering, so following this patter, no filtering options on
the request.

Closes #1615