feat: Create new logger
This adds a new logger builder which will create and return a new
`logr.Logger` which can be used throughout the codebase.

Essentially meaning that, this time, when you set all logs to be off
(or `"fatal"`), there are legitimately no logs.

It takes 2 configuration options:
- `level`: this corresponds to a string flag on the `gitops-server`.
- `devMove`: this is inferred from the server flag `--insecure`.
  Basically if you are not setting TLS certs it is assumed you are in
  dev mode and the logs are configured differently. Alternatively I can
  set up a dedicated flag or env var for this, so let me know.

I chose to stick with the logr zap wrapper since it is already known in
the codecase, and other dependencies like the ctrlr.Runtime use it,
meaning it can be seamlessly transferred to each component.