feat: Create a namespace client and watcher
Tried to use the Watch API, everything worked as expected, but I
couldn't get `testutils` in a good shape because Manager in controller
runtime does not expose WithWatch clients.

Controller-runtime has a separate Client and WithWatch (which is a
client with an extra function). That wouldn't be a problem as they are
basically the same, I tried to update `client.Client` everywhere to be
`client.WithWatch`, and the last piece was in `pkg/testutils/testutils.go`
where we are using Manager that has a `GetClient()` function, but does
not have a single function that would return a `WithWatch`.

Because of that, we are using polling instead. Maybe later we can
revisit this gap, for example if we have a wrapper client around
controller-runtime client and that can handle both `WithWatch` and `Client`.

Closes #1724