Upload docs to staging.go.w.w/branchname
This means we can have multiple PRs that affect the docs site open at
the same time, without them clobbering each other.

This also adds a status check to the github checks that take you
straight to the staged website.

Finally, it turns off the gcp cache for the uploaded docs, because the
whole point is that we want to be able to change them - the default is
a 1 hour cache.