Add a date prefix to the docker tag for PR images
We want to use Flux's image policies to auto deploy the latest version
of main to a staging cluster. To do this the images need to be tagged
in a sortable manner. The easiest way to do this is using the unix time
so we'll add that as a prefix.