Share tag between both images
Use a single, pre-generated, tag for both docker images (gitops and
gitops-share).

As well as making it clear that both images are from the same build (as
they'll have the same timestamp and sha)) this works around a limitation
in github actions: outputs from matrix jobs overwrite one-another (at
least according to this thread:
https://github.community/t/bug-jobs-output-should-return-a-list-for-a-matrix-job/128626)

In theory this could be worked around using the same upload/download
system as is used for the docker image but that feels like overkill.

The reason the tag is needed for the upload step is so only it is
uploaded (e.g. no `latest` etc).