feat: Add over-engineered reboot script (#1804)
Well this one got away from me.

I have paired with a couple of people recently and given them my "turn
it off and on again script" and they were suitably impressed that I
thought I would refactor that script and bequeath it to the masses.

As what often happens when I write bash, it went a bit overboard. So now
I present to you: The Complete CLI experience:

```
Usage: ./tools/reboot.sh <COMMAND> <OPTIONS>

Script to tear down and reprovision a dev environment for weave-gitops.

WARN: will destroy and recreate the kind cluster.

Note: this script is deliberately bare-bones. It will not be extended.

ENV VARS:
  GITHUB_TOKEN              With repo scope. Required for `bootstrap` command.
  KIND_CLUSTER_NAME         Set to configure created kind cluster name. default: wego-dev

COMMANDS:
  install                  Install flux to the kind cluster. This is the default action. If you run `reboot` without any commands, install will just run.
    OPTIONS:
      --namespace, -n      Namespace to install flux to. default: flux-system

  bootstrap                Bootstrap flux in the kind cluster and connect to a given gitops repo
    OPTIONS:
      --git-provider, -p   Git provider. default: github
      --owner, -o          Owner of gitops repo
      --repo, -r           Name of gitops repo
```

If left untouched, the default command `./tools/reboot.sh` will:
- Delete the `wego-dev` kind cluster
- Create the `wego-dev` kind cluster
- Install flux to the `flux-system` namespace

To bootstrap the cluster with flux and a gitops repo, run
`./tools/reboot.sh` with the correct options.