Add a fallback option to output the latest tag (#1834)
This is available either by running `tag-release current` or just
`tag-release` with no arguments. I realized I wanted some way to ask
"hey, where are we at right now?", and figured that this was a
reasonable way to do it. It also seems more approachable than spitting
out an error when no arguments are provided, which feels like a useful
property to have in a potentially dangerous script.