feat: Derive Cluster from Source object response (#1862)
The final UI piece for Mutli-Cluster #1677 (I think).

Previously in the UI we hardcoded "Default" everywhere in the Source
table and source Info pages. Now the UI will derive the object's Cluster
from its server response and fill the table/fields with that.

I have also added a Cluster field in the HelmChart source info page,
since we did not have that before and I wondered if it had been missed?