Make improvements to update-docs script
For updating download link, it does a greedy match, and apparently this
means it breaks the download instructions every release. Fix the
script, not the outcome.

For printing the command output, it apparently starts eating the
surrounding documentation every time you change the number of lines in
`gitops version`. I decided to close my eyes, hold my nose, and just
change the number. Oh, and I removed the deletion that's on the next
line, because it seems to work?

Stop overwriting the category json file for the CLI reference - it was
set to 8 intentionally, and we reverted it by accident. Just use the
file that's already there.