Move javascript out of the server binary
The server binary now just loads files from disk, instead of
files embedded into the binary.

This has the advantage that the build process becomes slightly more
straightforward - we can compile the go binary without building
javascript, and we can refresh the javascript in docker images without
re-building the go binary.

I've hard-coded the server binary to look at the sub-directory
"dist" below the binary's directory, and dumped the generated files in
`bin/dist` - it's a bit of an odd place to put the dist files, but
`bin` is a directory of generated files already, and this makes it
very easy to accidentally get right.