Stop exposing the Applications endpoints in core
These are all related to auth, and core doesn't do auth, so let's
disable them here for now.

These are still used in enterprise, so this doesn't delete the
code. Deciding whether it should be moved away or not is a future
problem - this at least shrinks the exposed surface running core.