ci: Redo snyk integration, again
Throw out docker integration - it's not helpful when all the binaries
we start are statically linked.

Don't use github security. Snyk-delta supposedly does the same thing -
let's see.

Finally, have `snyk monitor` break the merge pipeline if the error
code says that something went wrong with snyk, but not if it
successfully found problems. This might act unpredictable when there's
both errors and problems at the same time, but let's avoid that
scenario.