docs: Add helm chart reference
This uses [helm-docs](https://github.com/norwoodj/helm-docs) to
generate a reference out of the values.yaml - the only thing you need
to do is to add `--` to the beginning of every block of comments that
is supposed to be used for the manual.

I decided to create a new directory `references` in the manual, and
store both the CLI reference and this helm reference next to each
other on equal footing.

I have not documented every single option, so there's a lot that just
say "type: object, default: {}, description: ''" which isn't ideal,
but those are mostly generic, auto-generated properties which I
frankly don't even know what they do.

This file will be automatically re-generated for the `main` version
whenever we make any type of release. When we tag a release, this will
be re-generated before we fork the documentation, so both main and the
forked documentation should contain the newest instructions.

In case something goes horribly wrong, I've added a comment at the top
spelling out which helm version we generated the reference for.

This fixes #2002