Stop using logrus
We were apparently using logrus to
 a) print that there is a new version of gitops available
 b) error & exit if viper failed to the point where we can't start
 c) set verbose logs

Now, there's no logs to make verbose, so remove that. Error & exit is
perfectly supported by the built-in log module. And as for printing
that there's a new version available, we can just use the same custom
log framework every other log line in this binary uses.

Logrus is still an indirect dependency, but at least our code is more
consistent.