Put tsconfig & parcelrc into UI build dockerfile
I _think_ this might have been the cause of the frontend being
compiled differently when built with the docker file compared to when
doing it outside - tsconfig specifies flags that sound like they could
affect the compiler output.

A reasonable response to this would be "should we stop trying to
enumerate all the files, because this sounds fragile?" to which my
response would be, yes, but, I really don't want to bust the js cache
because go code changed, so, no, I'm gonna continue doing it this way
for now.