ui: Set kind property for sources, too
In 694dee7cfdf7a7b52765a4096a231457cfaaab4c, we changed the Source
interface to have a kind property instead of a type property. However,
the hook wasn't updated to set the kind property, instead it still set
the old type property.

This meant that the SourceDetail component, which finds the requested
source by name and kind, never found anything because it looked for
the wrong property.

This makes us set the kind property in the first place, which makes
SourceDetail work.