Fix specifying namespaces on sources
There are 2 types of automations: helmreleases and
kustomizations. They work differently when it comes to sources.

The kustomization's source is specified in a sourceref. The
helmrelease's "source" is always an auto-generated helmchart, and that
helmchart in turn has a sourceref. Make sure that we read the
namespace from the right property in AutomationsTable.

The sourceref is relative to the automation - that is, if it's in the
same namespace, then the sourceref's namespace is empty. I don't want
to care about this on the frontend, so I decided to always set it in
the API.

The helmchart doesn't have its own namespace - you couldn't set a
namespace, because helmchart objects are created by flux, not by you.
Through manual experimentation, I've determined that the helmchart is
created in the same namespace as the source, whether that is the same
as the helmrelease or not.