Start building a k8s-to-pb converter test suite
The parent commit made sure that the source namespace is set
and correct, whether the source is in the same namespace as the
automation or not.

This adds a bunch of tests, mainly focussed on that sort of thing,
but I want to put more and more stuff in here to part regression test,
part document how certain input fields are handled.