docker: Use cache mounts when building gitops-server
When building locally, this means the go cache will stay hot, and
builds will be a lot faster.

I don't know how much the parcel cache mount help, but if it's worth
caching, it's probably worth keeping that cache for next time?

When not building locally, it shouldn't make any difference.