ui: Put clusterName in all source links
It's absolutely possible for the same name/namespace source to have
different definitions in different cluster, so make sure we pull the
definition from the right one.

More importantly, now that this parameter is here, we can write an
endpoint that lets us just look up the correct source in the correct
cluster, rather than querying everything and filtering the results.