Add new endpoint to return raw object data
We need the full object to be able to display objects as yaml.

I could add this copy of the object to every response object, but it'd
be quite redundant - this response already contains everything that
the regular responses contain and more, so if anything I'd rather use
this object more rather than the hand-written API we use already.