ui: Put a yaml view + metadata view on detail pages
This is incredibly poorly styled, but it's put something into place
that we can iterate on.

It slightly reorders the AutomationDetail page so that the hooks are
all pulled in at the beginning - I was getting errors about us
violating the laws of hooks, and I think it was because of the early
return at the beginning.