Remove npm preinstall target, because it did odd things
`npm ci` was actually modifying the package-lock instead of
erroring. As a result, we've added specific step to our CI to manually
error out. However, the modifications `npm ci` was doing caused
dependencies we actually needed to be removed - so my options were to
have an invalid package-lock that wouldn't build, or have a valid
package-lock that was rejected because it wasn't up to date.

But why is `npm ci` modifying package-lock? Isn't the whole point that
it's the command that rejects non-up-to-date package-lock, that's what
makes it CI appropriate? Just remove the `preinstall`, and it actually
works.