js: Test that all dependencies are listed, add yaml dependency
It seems I didn't add yaml to package.json - but everything worked
anyway, until we tried to publish a npm package, and that was the
point at which it broke. Let's run the command that discovered a
problem on every PR, to make sure I don't screw this up again.